Main, StartEvent, StopEvent.

+++ src/Lab5/CarWash/State/Car.java
@@ -4,21 +4,21 @@ import Lab5.Simulator.Event.Event;
 
 public class Car {
 	private int id;
-	private int arriveTime;
+	private double arriveTime;
 
 	public Car(int id) {
 		this.id = id;
 	}
-	
+
 	public int getId(){
 		return this.id;
 	}
 
-	public int getQueueingTime(Event e) {
+	public double getQueueingTime(Event e) {
 		return e.getTime() - this.arriveTime;
 	}
-	
+
 	public void setArriveTime(int time){
-		this.arriveTime = time; 
+		this.arriveTime = time;
 	}
 }
